chore: modify create children

+++ src/main/java/app/mobile/BK_Kindergarten/children/ChildrenController.java
@@ -44,7 +44,7 @@ public class ChildrenController {
      * GET /children/{id}
      */
     @GetMapping("/{id}")
-    public Children getChildren(@PathVariable Long id) {
+    public Children getChildren(@PathVariable Integer id) {
         return this.childrenService.getChildrenById(id);
     }
 }
